refactor: split item sidebar count helpers
